requirements.txt
asgiref==3.5.0
autopep8==1.6.0
black==22.1.0
click==8.0.4
colorama==0.4.4
Django==4.0.3
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
tzdata==2021.5
